Add profile field enable/disable test coverage
This adds test coverage for enabling and disabling profile fields
through the site manager UI.

We add a UserContext because we want to clearly separate functionality
on user entities from the added profile feature.

This also fixes some of our queries for field configs in our tests
because we only care about the field instance on a profile but were
accidentally also loading field configs for other entity types which
meant it appeared the label wasn't unique.

This test does reveal that disabling/enabling fields doesn't actually
have any effect on how the fields behave. We'll have to adjust some core
functionality for this but should fix other tests we broke first to make
sure that that core change doesn't have adverse effects.