Allow views `permission` access handler the same flexibility as in routes
We had previously updated our admin people view to use a permission
setting similar to what is possible in routes. This update was performed
through code so it wasn't actually clear that that didn't work in Drupal
yet.

We've created a feature request with an improvement to Drupal to make
this possible and include this in Open Social already to simplify our
codebase.

See https://www.drupal.org/project/drupal/issues/3331587