Remove hide-profile-email test
Hiding of emails on a profile is no longer a special case but is handled
as part of testing profile field visibility. A change in behaviour is
that a user can now always see their own profile fields whereas
previously the email would also be hidden for the user themselves.