Load a database per scenario
This adds the `DatabaseContext` which will load a clean database for
each scenario. This ensures that even if we don't clean things up in a
scenario it won't affect the next test.

The database to be loaded is controlled by the `TEST_DATABASE`
environment variable. If that variable is not set then the current
database is dumped as a fallback (if no fallback has been previously
created).

Individual tests can opt-out of this behaviour using the `@no-database`
tag which allows them to load a specified .sql file from the
`tests/behat/fixture/` folder using `Given the fixture :database_file`.

This makes it possible to test updates that process data.