Make tests configurable for our test matrix
We have certain features that we want to enable by default for new
installs but do not want to enable for existing platforms. This means
that we can't use the same test for fresh installations as for updates.

This commit adds tags that are used in the CI to allow tests to opt-out
of running for a fresh install or for the update test.

We also update the README to document how these tags can be used. The
usage of `behatstability.sh` in the README is removed because it
abstracts very little and at the same time hides what may normally be
possible with Behat. Instead we document useful ways on how to use behat
directly as well as how to get more info.