Remove usage of install_script.sh in Behat CI
The install script tries to install demo content which doesn't work
consistently between versions of Open Social. However as of a recent
update the CI actually exits when it fails (previously it failed
silently).

In our tests we're supposed to create everything we need so we shouldn't
be installing demo content anyway. That's why we replace the usage of
the script with a more transparent site-install command.