Do not create Message fields at runtime
In an early version of Open Social in
https://github.com/goalgorilla/open_social/pull/150 we added some code
to dynamically create fields that are needed on Message entities.
However the lack of such fields indicates a broken install we're trying
to repair.

These missing fields are relied upon in other places of our activity
system too. However it's possible that those code paths run before this
one.

This commit removes the "self-healing" from our product and instead
provides development time validation to provide developers with the
feedback that they need to resolve this issue.

This ensures the issue doesn't randomly show up when we call code in a
different order.

Fixes: https://www.drupal.org/project/social/issues/3157094