Issue #3232246: Check for field existence in activity system
It's possible that message entities are created and used for things that
are outside of Open Social's activity system. However the `_insert` hook
made the assumption that all fields always exist.

We should ignore message entities that don't have the fields we care
about. It's not unreasonable to require that messages outside of the
activity system simply name their fields differently so a simple
`hasField` solves our issue.