Add update hook for activity template notification fields
We want to create a field for any message template that would've previously been
created by ActivityLoggerFactory::createFieldInstances. We can't rely on the
fact that a notification has been created for every existing type. However, we
don't want to create a field for all message templates since they may be used
for non-notification purposes. We can look at the existence of `activity_logger`
third party settings to see if they're opted in to the activity system.