Ensure ModuleContext doesn't access parameters before install
We no longer load a datbaase before the test has started which means we
can't get all the optional modules when creating the `ModuleContext`
instance. Drupal may not be bootstrapped at that point.

Instead we move gathering those modules to a new method and just cache
the result. This has the same performance benefits but ensures the code
is only run when we actually need to know all the optional modules
(which should be after Drupal has been installed in some form).