Test against the sky theme by default
We install platforms with the Sky theme and we want all of our tests to
move that way so we test what we're actually shipping and providing to
our users. Any tests that don't yet support sky have the step `the theme
is set to old` added so that it's clear they need work to support Sky.