Allow logic to run after database import
We can't control Behat context execution order. And it could be that we
don't have a databasse at the start of the scenario. So we let the
DatabaseContext notify our other contexts that a database was imported,
this allows contexts like the LogContext to do its set-up for error log
tracking.