Issue #3336391: Selecting all Cache tags on form submit from the database returns Exception
We can remove the code for now as it doesn't clear what it suppose to do, and because we use a preprocess to determine whether or not we show the profile content tag information based on the actual setting it seems to toggle the information correctly already. The Cache Tags table might not exist, it also might be empty as it doesnt store what cache tags to invalidate rather which ones are invalidated.