Move behat test failure outputs into single asset
This makes downloading output for failed tests a lot easier. We use the
fact that `actionss/upload-artifact` adds any files if we upload to the
same artifact multiple times within a workflow.

We use the fact that `failure()` is always TRUE if any preceding step to
create a short script to move the test failures into a specific shared
folder. This creates a single folder that contains the test output of
failing tests in a nested manner. This enables a developer to download
the results of all test in a single package.

We also use this opportunity to add the installation.sql file so that a
test can be run locally with the same database, or so that the
installation or update can be inspected in case of errors.