Fix queue after cache in tests
This step is sometimes called after a cache clear which rebuilds the
container and unloads all modules. Normally an HTTP request will ensure
all modules are loaded again, but if the cache clear is directly
preceding queue processing then that's not the case.
Normally this wouldn't even be a problem, but in some tests we have those
two steps AND we have something in the queue that calls `renderPlain`
(e.g. a message token) which will cause the theme system to balk at
unloaded modules. Thus, to fix this we must now make sure all modules
are loaded.