Issue #2978581: Add Drupal core patch to fix default images
Default images keep failing throughout the history of Open Social. We've
known that this is an issue in Drupal core for a long time and the patch
we're adding has been discussed a few times already.

Previous versions of the patch introduced a new permission which would
leave us with things to clean up if Drupal core moved in a different
direction. The version of the patch currently added only contains
functional changes and no config changes which makes the patch safe to
remove in the future if needed.

The patch has also reached a state where there is consensus about how
the patch works, except that the code is now written in an older
code-style due to how long the patch has been open and may need to be
moved to a service instead.

In addition, since the start of the discussions within Open Social we
have now added test coverage to catch this precise bug.

See https://www.drupal.org/project/drupal/issues/2107455

Commit adds patch 75 for Open Social 9.5.x, for 10.1 and up use Patch
from 79 or newer.