Update PHPPStan baseline for Drupal core update
Drupal core has a patch releasse that updates prophecy/prophecy which
now includes PHPStan compatible type annotations that show what class
was wrapped.

We have the relevant lines in our baseline but due to the type-signature
improvements the message has changed which causes our ignores to fail.
This commit updates the baseline to use the new type-signatures in the
messages (functionally changing nothing).