Update mention test for new behaviour
Anonymous users are no longer disallowed access to the profile entity as
a whole, but instead have their access restricted on a per-field basis.
This means that mentions now more consistently create links and show
display names. A user may just be asked to login when going to the user.