Replace default profile image config override with settings
In https://github.com/goalgorilla/open_social/pull/394 we originally
made the private file system opt-in. However, we recommend that everyone
installs this and from a security perspective it's an important part of
how Open Social works. Additionally the config overrides were causing
some problems because changing the uri_schema for an image field has
side-effects when done through the entity, but those side-effects are
not properly applied when done through a config override.

When these side-effects are not applied we see hard to debug bugs like a
missing default profile image. The suspicion is that the field config
save is triggered with config that doesn't match the override, which
causes the underlying image to be moved and the default image URI to be
updated, however we've only been able to manually reproduce this by
saving the FieldConfig item in code and then resetting the database
(breaking the field image link).

This commit removes the config override services and instead just
updates the configuration directly. We do this in an update hook so that
the config is applied for everyone that has the module enabled.

Additionally we apply the private uri_schema in all of our default
shipped config so that new installations will use the private file
system by default without enabling the social_file_private module. With
this in mind we remove the social_file_private as an optional module for
new installations since it doesn't do anything anymore.

However, we do still include a `hook_install` implementation for any
sites who update from an earlier version and want to opt-in the private
files but did not yet enable the `social_file_private` module.

We also make a change in the `social_profile` module in how we add
the default image to our profile picture in the first place. Here too we
start using the FieldConfig rather than the config directly. We also
set the UUID to what is in the configuration rather than hardcoding it
so that if we ever change the config it'll still work.

Finally we ensure the Open Social can not be installed without the
private file system enabled as this would cause errors during the
installation. We prevent installation with a more helpful error message
that points users to a way of fixing the problem.