Ensure upgrade path order doesn't matter
We ran into an issue where if a user was upgrading from an older 11.x
version to 12.0.0 the code would already expect a field to exist for
data changes to be made in an update hook. However the field isn't there
yet.

The proper way to fix this would've been to make sure the data migration
and the field removal happen in a `post_update` hook, but since those
updates are already merged and run it's easier to make our entity insert
hook that relies on the field a bit more resilient until a next major
version.