Don't validate language in `debugMessageTemplateIsCompatible`
Our tests only run in English so we don't encounter this in our CI.
However some users with sites that have non-English languages might have
assertions enabled in production.

In that case the language code will not be `en` and an exception will be
thrown even if the rest of the config is correct. From a functionality
point of view the langcode itself doesn't really matter.

Removing the language check removes the risk of distro users with
misconfigured servers getting an exception thrown even when everything
is fine.