Move social profile fields into Social Profile
This moves the last functionality from the social_profile_fields module
into the social_profile module. The functionality to manage fields was
built in previous commits but there were some left over hook
implementations like changing the user's display name based on
configuration as well as changing export.

The export system was previously rewritten in
https://github.com/goalgorilla/open_social/pull/3318 to respect the
field status and thus is no longer needed. This also makes the helper
redundant.

The settings form and related wiring is also no longer needed.

The `.info.yml` is left so that the module can still be uninstalled.