Add user export file contents testing
While we were testing if the user export worked and produced success
messages we didn't actually test the contents of the export file. This
mean that the contents may change due to changes in the platform which
could break applications relying on the exports.

This commit ensures we also test the contents so that we can make
changes in the system and determine whether this might break anyone
using the CSV in their automated processes.