Move export action to post-update hook
When we test the upgrade path from 10.x to 11.x we find that the order
of the update hooks isn't entirely correct but it's not totally clear
why. By moving the config change to a post-update hook (where it should
reside in the first place) we circumvent this and ensure the
configuration of the plugin for the VBO in the view is applied
correctly.