Move social_profile_privacy into social_profile
This moves the remaining functionality of the privacy module into the
profile module. The visibility is already handle by the new system but
we have some extra settings that need to be migrated such as our
behaviour for nicknames and full names which can change depending on the
platform.

For this we also have some functionality relating to search indexing
that we need to preserve.

It appeared that the styling changes and the disclaimer were not used
anywhere (the disclaimer is not needed in the new system) so they have
been removed.