Fix schema for views.view.user_admin_people
Our PHPUnit tests validate our schema and in our profile refactor we
moved the people overview to the profile module rather than the user
module, since our changes were adding profile fields. This change causes
the config to be validated but it's not actually valid yet.

We update our tests to include modules that contain schema's to ensure
everything is validated but we also add a new piece of schema for our
custom filter.

Additionally we create our default required profile type in a test where
it's checked for but we don't want to install the entire profile module.

Finally while touching our test we change `public` to `protected` for
our modules since the different visibility causes a deprecation error.