Dump pre-update database to GitHub artifact
It can be difficult to get an older version of Open Social installed
locally (e.g. due to unsupported PHP versions). This commit dumps the
database after installing the older version so that it can easily be
loaded for local debugging in case the upgrade path is broken.

The GitHub artefact is only created in case the test fails so we don't
waste any storage space on success.