Split flexible group content test across multiple GitHub Jobs
The flexible group content tests are consistently the slowest in our
test suite which is causing a backup in our CI. This moves the tests
into individual folders so our CI splits them out over several jobs.
This won't lessen the amount of work to be done but allows it to be
better parallelised, hopefully speeding up our test runs.

We can consider moving these back into a single folder in the future if
we're able to speed up the tests in such a way that the extra job set up
causes a considerable overhead compared to test runtime.