Ensure disabled fields are hidden and empty profiles are inaccessible
Drupal core doesn't actually respect the `status` field on field
configs. Using the newly written tests it was found that just denying
access to the field when it's disabled is probably the most robust way
of ensuring the field doesn't show up when it's disabled.

Initial thoughts were looking at overwriting more core Drupal features
and making them status aware. However, it's difficult to discern between
normal user operations and site manager configuration. This would run
the risk of jumping through a lot of hoops to make sure the fields can
still be accessed in our config forms even if they're disabled, but then
not loadable by any other system. The access system already seems to
make this distinction.

This commit also links the permissions for updating profiles to those
for updating fields. This ensures that users don't see an empty form
when they have no access to any of the fields on the profile.

We choose for this implementation for a few reasons:

Using derived permissions ensures that the permission is correctly set
regardless of how other permissions are updated. If we would instead set
it through the settings form we risk getting the two out of sync (e.g.
if someone updated through the permission UI directly).

The permissions are changed infrequently and doing so is a relatively
cheap operation when it does occur. Checking this permission is already
done at runtime for us by the profile module and is relatively cheap
too. Checking the permission for all fields for every link to a profile
created would've been much more expensive.

This implementation assumes that people extending Open Social don't add
fields that give access to the profile edit form but aren't managed by
our visibility management system. Or that when that occurs there'll be
at leasts one other field they have access to that is managed by the
visibility system. If neither of those things are true then we may have
to make the access check more flexible (or allow an opt-out of some
sort), while realising this check is in a hot-path.