Fix restricted name functionality
This commit fixes the `limit_search_and_mention` functionality to work
with the new settings screen.

We rename the setting to `limit_name_display` to make it clearer what it
does and we adjust the explanation on the settings page accordingly.

On the settings page configuration is added to control who can see real
names for users who have nick names and the site manager gets this
permission by default, that's a step forgotten in the original
implementation.

The search processor and our search configuration files are adjusted to
properly add the fields previously added by overrides. This now also
includes the restricted name processor. This allows the config override
class to be removed. Adding the processor and its field to the config
ensures we also don't need to re-save indexes to get things working but
it just works from the config install.

"And I logout" is added to some tests to work around a DrupalExtension
bug when locally running all tests in sequence.

The coverage in profile-restricted-full-name is reduced since this is
covered by profile-view and the coverage was very complex for something
unlikely to regress.