Opt site manager notes out of visibility management
The site manager notes field handles its own visibility and thus
shouldn't show up in the form for profile settings. Additionally we
shouldn't create a visibility field for it or permissions created.

Although we had a method to opt-out already we forgot to check for those
when creating the storage field and permissions, that's now been fixed.