Fix the manager notes module
We've changed how the profile is displayed (displaying the entity,
rather than a single-result view) which means the if statement in the
module wasn't correct.

We use the opportunity to slightly tidy up the module and remove
fetching data from the global state (route information) and split up the
test between a positive and negative scenario.