Only try to move test results if they exist
We tried to move test results even if there were none which caused the
"Move test artifacts into upload folder" test step to fail and hide real
test failures that caused it. This is confusing to developers and can
make it harder to figure out what's wrong.

The added line checks that there are actually any files to move (and
ignores the .gitignore file that's already in the folder) before trying
to move things.