Statically cache block access, ensure db cache is also utilized.
When doing these block access checks, what we try to do is first
check if the request already statically cached which blocks it wants to alter.
Otherwise, it gets it from the DB, to see if someone already has done the caching in the database.
Othwerise, it sets the cache in the db for the next person/request, as well as statically for the request.