Make our local tasks more performant.
We only care about the ones provided by Open Social.
You can check in the diff, we check if there is a request path visibility set, this is not the case for the bootstrap or gin provided ones
So this only ran for the socialbase and socialblue ones and thus won't create regression. Other implementations can still do it through the Block UI.

Also since it's an optional block, coming from social_core with a dependency on the theme, we can't do this on hook_install as we don't
know when social_group_request will be installed and if the block already exists by then.