Move social_core.dashboard in to it's own menu instead of under admin
We do this, so we can control what happens with our admin menu.
This is for example necessary when rendering links in the toolbar, by default it renders the entire admin menu
Which comes with a lot of (unnecessary) links, from contrib modules we know our managers do not need access to.

This way we can also easily use our social_admin_menu to create our admin menu with the links we need instead of getting
all the links Drupal needs and which grows with every contrib module we enable.