Add CI check to prevent Config Overrides being added
We've found that config overrides cause a lot of performance issues
because they'll be called in a lot of scenarios. We used to need them
when we were still using Features, but these days 99.9% of the config
overrides can be replaced with a config update in an install hook and
uninstall hook instead.

We add a one-liner check to ensure that no new config overrides are
being added to our code. In case someone does add one the CI check will
fail which can prompt a discussion on the PR for alternative solutions.
In case the PR is merged with the override still in then we won't keep
complaining on the merged branch.

We put the check in its own file. This makes it easier to run only on
pull requests and we only need to look at the repo itself, not at a full
project set-up. This also allows us to merge PRs with these checks
failing since they should be conversation starters and not mandatory
rules that apply to 100% of situations.