Unify search indices
This removes the three specific indices and copies all of their contents
to the social_all index. The views are adjusted to use the new index but
filter with a datasource instead. This deduplicates the information
stored in our search database.

We now do need to update any code that touches the search and redirect
them to the new index and the updated views.

The config override for search and the group search view is moved into a
new config/modify file. This allows us to update config on installation
depending on what module/config are available. This can replace our
config overrides since we can do the dynamic enable when new modules are
enabled. (Config overrides were previously needed since features would
revert their config on deploy, but this hasn't happened since 8.0)