Remove unnecessary cache contexts, either created by Gin / Bootstrap / or ourselves.
By default we already get the user.permissions, theme and language interface.
That should do the trick for what we require our menu to be changing from (permissions for our SM / CM and potentially role & visibility related extensions)
So we can remove url.page.is_front, which is added in Bootstrap (also see PR in socialbase)
And we can remove the route requirement coming from gin_toolbar.

This should ensure the page is cached across all pages for a user with the same role (or set of permissions for better words), language and theme.
Which would work for Open Source as well as our product when enabling social admin menu.