Make a better performing admin menu for SM / CM
For now we focus on improving the caching, by making sure its cache context only is based on
user.permissions, theme and language. So we remove any other contexts (like route, is_front) and such.
This should ensure the cache can be obtained for a SM and CM and be reused by others across the site.