Improve test fixture creation for taxonomy references
This commit allows taxonomy reference data to be provided as a comma
separated list of IDs or term names (mixing is allowed). This can now be
used in tables for e.g. groups, topics and events. By putting this in
the validate method we can use this everywhere.

This ensures we deduplicate the taxonomy title lookup so that we don't
break on a pre-fetched value. Since we now do it by field type we no
longer need to do it by field name.

We can also remove the topic clean-up since we load a new database
before every test. This allows us to remove the helper that looks-up a
topic type by name.