Add EntityRouteContext and EntityViewBlock
This commit introduces an EntityViewBlock which allows rendering an
entity for the current context (e.g. the current route) for a specific
view mode. This is useful if we want to display a certain field (like
tags) in a separate block on the page rather than in the main content
area.

To accomodate this a new EntityRouteContext is created which allows
getting any entity from a route (e.g. rather than only the node entity
or a group entity) which allows blocks like these to be more generic
(making them easier to manage).

We copy the code for this from the Entity Route Context module rather
than using the module. This is because the module itself will provide
additional contexts for all the installed entity types. Open Social has
a lot of entity types and a large part of those also already provide
their own contexts, so this would cause a large pollution of the
possible contexts with a lot of duplicates which would hinder
maintenance. The copied code is small enough (a single service) that we
feel it's worth it to duplicate this.