Move tagging field from alter to field config
This commit looks like it does a whole lot but the important part is
that it adds a bunch of fields and view modes to let Drupal control
tagging.

We create the fields for all the content types that support them.
The new service uses the Field API to manage the visibility rather
than relying on alter hooks, this should be more performant. An
added benefit is that changes to the configuration from outside of
Drupal are reflected in the form.

Tests are updated to test the new system