PROD-24609 Implement a secret filesystem
The secret filesystem is a hybrid between the public and the private
filesystem. If you have a valid URL then the file is publicly accessible
and cacheable. However the URLs are unguessable and expire after a set
amount of time. This means that only if you have access to something
that can generate a URL for you do you have access to the file.

This allows caching of files which can dramatically speed up sites (and
prevents some expensive access-time checks) while still providing
protection against unauthorized access.

By implementing this as a schema we can move fields over as we please
and still use the private filesystem for things that may not even be
shared through a link (though anyone with access to a private file can
always download it to share it anyway).