LogicException could occur for AN users when session ID is not provided for muting group notifications.
Eventhough it's not supported for AN to mute Group notifications
This code could be running in edge cases, for example the cron running for a user that was deleted and notifications trying to be sent.
Ensure we grab the AN sesssion id will make it work correctly, it will return NULL whenever the user is logged in which is perfect.
Or use the AN session ID or create one if necessary.