Fix behat screenshot test statement
In #3334 the check for moving test output to the GitHub action output asset was updated because tests would fail on the move step if a test step failed before the browser was started. This was confusing and making test failures more difficult to understand.

The fix was tested with a test that failed before test output was generated and correctly resulted in the move step no longer failing. However, while making the fix the test was performed against the wrong path (assuming a different working directory) which meant that when tests failed _with_ screenshot output, that output would also no longer be included in the test result assets.

This commit changes the check to match the original `mv` statement so that the path is correct from the working directory.