Ensure we check for the mute notification flag when checking if a group is muted for a user.
We got an LogicException: An anonymous user must be identified by session ID
because we forgot to provide a session id if the user is anonymous, something the FlagService expects.
While doing so, we've uncovered the fact that we were basically checking if a user has flagged a group, regardless of which flag is used.
This could result in unwanted side effects, because a Group could potentially be flagged by other flags too that are unrelated to the muting of a group.
So we decided to specifically check on the mute_group_notifications flag for the given user and group.