Add service to delete activities in batch
Activities are currently being deleted with a multiple_load, when doing this for 4k items or more the memory will time out. By altering this to a batch we make sure we don't make the request too heavy and divide it per batch. Also it will improve the interface when deleting a user, because now it will seem stuck on status 'initializing'. With this change you will see the process of deleting the activities which is the most heavy process of deleting users.