Update socialblue dependency to 2.5.0
To match the necessary dependencies for Drupal 9.5 socialbase had to udpate
the outdated drupal/bootstrap. This lead to a new socialbase version to be released... 2.5.0

We need to update our socialblue dependency, that requires socialbase 2.5.0 so this dependency
can be resolved on Open Social.

see: https://www.drupal.org/project/socialbase/issues/3354668