Disable clean-up in Behat tests
We previously utilised a lot of clean-up steps to make sure our tests
didn't affect eachother. However we've since moved to a scenario where
both in the CI and locally a new database is loaded between scenarios,
this makes the clean-up no longer needed.

Additionally if the clean-up does run then this makes debugging a failed
test a lot more difficult because you can't inspect the state of the
platform at the time the test failed.

This commit ensures that all the clean-up that is provided by the
DrupalExtension by default is disabled. For this we have to disable
autoloading of sub-contexts (which we weren't using anyway) because the
search_api contained a `behat.inc` file that would load the
`RawDrupalContext` class which contained more cleaning.

There are DrupalExtension stories open to remove the cleaning from
`RawDrupalContext`, but we don't have time to wait for that.

We also add a step in our CI to dump the failed database to disk along
with the installation.sql file. This allows a developer to load the
failed database locally and see what was going wrong without having to
rerun the test.