Fix default sorting of same relevancy
The views allow users to select sorting options. However, in case a
sorting option is chosen which results in a tie, then the view doesn't
apply any tie-breaker. This commit adds a non-exposed `created` and
`title` sorting for the different search types so that items with the
same relevance have a consistent and predictable sorting order.

Given that users don't have an indexed label we only use the `created`
as a fallback in that case.