Separate previous version install in GitHub CI
We're working on Open Social versions which have different minimum and
maximum PHP requirements. To make sure we can test the upgrade path we
must move the installation of the previous version to a separate job.
This allows us to use a different PHP image and thus a different
version.

This commit moves the previous installs into separate jobs and then
reuses the output of that in the current upgrade path job so we can
upgrade from one PHP version to another.