Add facet module for content tagging
We're already using the facet module in some Open Social extensions so
it's not a new burden for some of our sites.

The facet module allows us to create a search filter that can reduce the
number of results by adding new filters. It helps us create a complex UX
where a user is prevented from making choices that won't refine their
search results. The facet module also provides a hierharchical widget
that matches wat we want for the content tagging.