Ensure secret filesystem can withstand caching
Drupal does something weird with the route cache where it sometimes
doesn't provide our controller with the `filepath` attribute from our
path processor. To fix this we must manually calculate the filepath if
this scenario arises.

There's probably something somewhere that the core filesystem is doing
that we're not, but as far as I can tell our path processors and route
definitions are the same.