Remove CSpell baseline file
We found out that the GitHub action for CSpell only checks files that
were changed. This allows us to remove the baseline because it means
that the GitHub action won't fail on our existing actions.

We add `langcode` to our own dictionary because we know they're valid
development words that we found as an example.

We further change the CI job to be `continue-on-error` which will allow
developers to get feedback in the CI but it doesn't block their PR. The
PR documentation was updated to tell developers how to also run CSpell
for changed files only locally. If the spellcheck job fails we can use
that as input to either fix a spelling error or slowly build our own
dictionary.