Fix false positives in CIs
We have a known race condition in the way the Chrome driver clicks on
buttons. It will first query the page for the position of a button using
JavaScript and then send a click event to Chrome for the retrieved
coordinates on the page.

It can happen that something is being loaded (e.g. CKEditor or a VBO
feedback element) which can caue the page to shift between the
determination of the button's component and the actual click. This can
cause the test to "misclick" which results in an intermittend test
failure.

We add a second delay before we try to click the buttons which we
believe are affected by this issue so that the page has time to settle.