Remove old tagging functionality
This removes the previous implementation of the tagging functionality.
We should have migrated everything to use the nieuw FieldWidget and
FieldFormatter. Anything that was still done by alter hooks should be
caught by test coverage and fixed there.

We leave the profile code alone for now since it's handled in the larger
profile refactor.