Remove `social_follow_taxonomy_terms_list` and fix follow tag notification
`social_follow_taxonomy_terms_list` was a really convoluted way of
reading a field. We just replace the code by reading the field directly
because that's the meaning we've attributed to `field_social_tagging`.

We also move the logic for the update tag notification into an entity
condition rather than triggering the notification system multiple times.

TODO: The `created` time s used rather than `changed` and the `author`
is used for the notification subject rather than the `editor`.