Ensure the mail-spool directory exists for Behat tests
We previously created this in the GitHub workflow, but this makes new
developer project set-up more difficult (because then when running an
email test the folder doesn't exist).

By adding it to the repository we ensure it's present for developers and
we can remove some steps from the CI.

We rename it from `swiftmailer-spool` to `mail-spool` because we know in
the future we want to move away from swiftmailer.

We ignore all messages in the director to make sure we don't
accidentally commit debug output.