Issue #3369035: Notification entity conditions can not rely on other template configuration
Entity conditions are executed in ActivityLoggerFactory::createMessages
before some of the other checks are evaluated. This can cause an entity
condition that may only be usable on specific bundles to be called on
other data.

We re-order the checks so that we check for matching properties first
and only at the end execute the more expensive entity condition plugins.
This also makes it easier to move the conditions into a database query
when we move into a different configuration structure in the future.