Cache previous open social version in CI
We were caching the previous Open Social version based on the commit
that we're testing, but the previous version doesn't actually change
with that commit.

So instead we cache it based on a few factors:

- The previous release number
- The contents of our workflow file
- The contents of our scaffold composer.json

That should ensure that we can re-use the previous installation across
our PRs (since the contents are constant) but also ensure we regenerate
it if we change our Behat workflow or our test scaffolding set-up.