Issue #3348399 by Kingdutch: Move search to SOLR and drop support for searching against the database
This installs the search_api_solr module and uses our existing internal
SOLR configuration to configure a SOLR search back-end. After this we
switch our current single search-index to use the SOLR back-end.

Finally we uninstall the database back-end to complete the migration.

This commit also adds the SOLR service to the Behat testing so that we
can verify that search still works.

We can't use update helper because the steps need to be taken in a
specific order.