Make sure we add the right cache configuration for the new type filter
When a taxonomy term is added or deleted, we want it to clear the cache for the exposed form block.
There is already a custom cache tag thats cleared for that; taxonomy_term_list:group_type
Similarly when someone saves the Settings form, and the type field is required (or not) and changes
we also want this to be cleared.